Add support for environment variables and aliases
If an 'environment' or 'aliases' tag is found in environment.yml,
'conda-env create' will create activate/deactivate scripts that configure
the environment with those values when activated, and remove them when
deactivated.

This is done by putting two simple '.sh/.bat' files in
$PREFIX/etc/conda/(de)?activate.d

These files then evalute the output from a python script that does the
actual handling of adding/removing environment variables and aliases.